Dockerized react folder
We'll probably have to come back and change some Docker stuff since it's not optimized for deployment but it runs just like the server one. You can run it like normal or check the README.Docker.md to see how to run it with Docker.

I also added some code in the vite.config.js to get the Docker stuff to work correctly