Added setup file for MySQL database and phpMyAdmin
To run just type "docker compose up" and go to localhost:8080. It'll ask for login credentials and the IP address the server is running on. To find the IP you can check the terminal for the db in the Docker desktop app or you could just use "docker inspect db-db-1" which will output all the information you need including the IP (near the bottom).