Navbar works better with the pet page
Before the navbar would push the pet stuff down since they were on the same layer. Now the navbar is on top of the pet's environment. Still needs css though since it's hard to actually see it with the backyard background and there's no style to it other than it being on top of the scene.