JWT authentication added
Added JSON web tokens that are generated when a user logs in then stored in local storage. You can use the @jwt_required() decorator to a flask function to prevent unauthorized users from accessing things like somebody else's profile or pet. The token also stores the userID so we can setup our routes to easily include them by just getting it from the token. For example, if we have a profile route that looks like /api/users/<int:user_id>/profile then in the frontend we can decode the token instead of having to fetch the userID separately.