Framework for basic pet interactions added
Application now checks if a pet has been generated for a user when they log in and creates one if it hasn't (only works for existing accounts that don't have a pet now so I'll need to add it to the registration function too). Users can now interact with their pet by clicking on it in the pet environment. Pet interactions are secured using JSON web tokens so user's can't change stats on anybody else's pet but their own.

Saving interactions isn't currently included but I'll get on that next as well as adding some more interactions and better logic like limiting the number of times you can raise a stat a day.