Interaction recording now working
Interactions are now saved to the PetInteractions table. This will make limiting the number of times a user can do certain actions in a day easier. Also fixed a bug where clicking on the pet at a certain point in it's walk cycle would make it stop walking (intended) but then start walking before the menu was dismissed (unintended). Next up for the virtual pet is to add some more complex logic like action limiters, stat degradation, and mood based on stats