Reward mechanics partially implemented
Added some code related to the frisbee reward. When clicked, the frisbee will fly off screen and the pet will follow it. The idea is to hopefully get it to the point where the pet will follow and bring it back before the frisbee resets. Also switched out the rope reward with a bell. The rope was supposed to trigger a tug of war with the pet and the user's mouse but it was too buggy. The bell just calls the pet to the front of the screen