Matching game works with images now
Set up the matching game to take images and made an architectural engineering game where the user has to match different architectural styles with the images. I also made the matching game a component that can just be imported and passed the matching dictionary so we don't need to copy and paste the code. Added the reward notification for the matching game too.